1) Moved some floats in Particle structure that don’t have a small dynamic range of values from FLOAT to float to save memory.  2) Removed sinkid variable from Particle structure and replaced with flag to indicate if inside sink or not.

+++ src/Tree/BruteForceTree.cpp
@@ -356,7 +356,7 @@ void BruteForceTree<ndim,ParticleType,TreeCell>::StockTreeProperties
       cell.N++;
 	  if (partdata[i].flags.check(active)) cell.Nactive++;
 	  cell.hmax = max(cell.hmax,partdata[i].h);
-	  cell.maxsound = max(cell.maxsound, partdata[i].sound);
+	  cell.maxsound = max(cell.maxsound, (FLOAT) partdata[i].sound);
 	  if (gravmask[partdata[i].ptype]) {
 		cell.m += partdata[i].m;
 		for (k=0; k<ndim; k++) cell.r[k] += partdata[i].m*partdata[i].r[k];
