+++ src/Tree/KDTree.cpp
@@ -787,7 +787,7 @@ void KDTree<ndim,ParticleType,TreeCell>::StockCellProperties
         cell.N++;
         if (partdata[i].flags.check(active)) cell.Nactive++;
         cell.hmax = max(cell.hmax,partdata[i].h);
-        cell.maxsound = max(cell.maxsound, partdata[i].sound);
+        cell.maxsound = max(cell.maxsound, (FLOAT) partdata[i].sound);
         if (gravmask[partdata[i].ptype]) {
           cell.m += partdata[i].m;
           for (k=0; k<ndim; k++) cell.r[k] += partdata[i].m*partdata[i].r[k];
